WIP: Move register method from user to auth service
