+++ Models/Consts/ApiNamesValues.cs
@@ -1,0 +1,9 @@
+namespace minimal_api.Models.Consts
+{
+    public class ApiNamesValues
+    {
+        public const string AuthApi = "Admins API";
+        public const string CarsApi = "Cars API";
+        public const string UserApi = "User API";
+    }
+}
