+++ Routing/AuthRouting.cs
@@ -1,5 +1,7 @@
-using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
+using minimal_api.Entities;
+using minimal_api.Extensions;
+using minimal_api.Models.Consts;
 using minimal_api.Models.DTOs;
 using minimal_api.Requests;
 using minimal_api.Services.Auth;
@@ -11,11 +13,15 @@ namespace minimal_api.Routing
     {
         public static IEndpointRouteBuilder MapAuthApi(this IEndpointRouteBuilder routes, IConfiguration configuration)
         {
+            routes.MapPost("/auth/register", ([FromServices] IAuthService service, [FromBody] User user) => AuthRequests.Register(service, user))
+                .WithValidator<User>()
+                .WithTags(ApiNamesValues.AuthApi);
+
             routes.MapPost("/auth/token", ([FromServices] IAuthService service, [FromServices] IUserService userService, [FromBody] UserLoginDto userLoginDto) => 
                         AuthRequests.GenerateToken(service, userService, userLoginDto))
                 .Produces<string>(StatusCodes.Status200OK)
                 .Produces(StatusCodes.Status404NotFound)
-                .WithTags("Auth API");
+                .WithTags(ApiNamesValues.AuthApi);
 
             return routes;
         }
