+++ Routing/UserRouting.cs
@@ -1,7 +1,5 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using minimal_api.Entities;
-using minimal_api.Extensions;
 using minimal_api.Models.Consts;
 using minimal_api.Requests;
 using minimal_api.Services.UserService;
@@ -13,20 +11,16 @@ namespace minimal_api.Routing
         public static IEndpointRouteBuilder MapUserApi(this IEndpointRouteBuilder routes)
         {
             routes.MapGet("/user", ([FromServices] IUserService service) => UserRequests.GetAllUsers(service))
-                .WithTags("User API");
+                .WithTags(ApiNamesValues.UserApi);
 
             routes.MapGet("/user/{id}", [Authorize(ConfiguredUserPoliciesValues.AdminsOnly)] ([FromServices] IUserService service, [FromQuery] Guid id) => UserRequests.GetUserById(service, id))
-                .WithTags("User API");
-
-            routes.MapPost("/user/register", ([FromServices] IUserService service, [FromBody] User user) => UserRequests.Register(service, user))
-                .WithValidator<User>()
-                .WithTags("User API");
+                .WithTags(ApiNamesValues.UserApi);
 
             routes.MapDelete("/user/{id}", [Authorize(ConfiguredUserPoliciesValues.AdminsOnly)] ([FromServices] IUserService service, [FromQuery] Guid id) => UserRequests.Delete(service, id))
-                .WithTags("User API");
+                .WithTags(ApiNamesValues.UserApi);
             
             routes.MapDelete("/user", [Authorize(ConfiguredUserPoliciesValues.AdminsOnly)] ([FromServices] IUserService service, [FromBody] List<Guid> ids) => UserRequests.DeleteMany(service, ids))
-                .WithTags("User API");
+                .WithTags(ApiNamesValues.UserApi);
 
             return routes;
         }
